Avoid powerpc64 sign and zero extension for ZEND_VM_SMART_BRANCH
Casting an unsigned int to an int requires powerpc64 to sign extend
either explicitly via an extsw instruction, or implicitly via an lwa
instruction. In either case we end up with an extra instruction in
the pipeline.

In this case we want to go further and declare this as an unsigned
long, because gcc gets confused and zero extends the int to a long
in some users of ZEND_VM_SMART_BRANCH.